# swagger.yaml
swagger: "2.0"
info:
  description: "This API is to be used during development of assignments for the QUT CAB230 Web Computing Unit. If you have any issues, please comment on Slack."
  version: "1.0.0"
  title: "CAB230 : Web Computing API"
host: "localhost:3001"
basePath: "/"
schemes:
- "http"
...
paths:
  /register:
    post:
      summary: "Registers a new user account"
      tags: 
        - Authentication
      parameters:
        - in: "body"
          name: "Request body"
          required: true
          schema:
            type: object
            properties:
              email: 
                type: string
                example: "demo-user@gmail.com"
              password: 
                type: string
                example: "demouserpassword"
      responses:
        201:
          description: Created
          schema:
            type: object
            properties:
              message:
                type: string
                example: "yay! you've successfully registered your user account :)"
        400:
          description: Bad Request
          schema:
            type: object
            properties:
              message:
                type: string
                example: "oops! It looks like that user already exists :("
  /login:
    post: 
      summary: "Login with an existing user account"
      tags: 
        - Authentication
      parameters:
        - in: "body"
          name: "Request body"
          required: true
          schema:
            type: object
            properties:
              email: 
                type: string
                example: "demo-user@gmail.com"
              password: 
                type: string
                example: "demouserpassword"
      responses:
        200:
          description: OK
          schema:
            type: object
            example: 
              access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjo1Nzg2LCJlbWFpbCI6ImRlbW8xMjM0QGdtYWlsLmNvbSJ9LCJpYXQiOjE1NTUzMjQ2MDIsImV4cCI6MTU1NTQxMTAwMn0.mo85iKuAQ_h8n8-6pkeNNs5CCnV0G8hScxhi_ml_Q6c"
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjo1Nzg2LCJlbWFpbCI6ImRlbW8xMjM0QGdtYWlsLmNvbSJ9LCJpYXQiOjE1NTUzMjQ2MDIsImV4cCI6MTU1NTQxMTAwMn0.mo85iKuAQ_h8n8-6pkeNNs5CCnV0G8hScxhi_ml_Q6c"
              token_type: "Bearer"
              expires_in: 86400
        401:
          description: Unauthorized
          schema:
            type: object
            properties:
              message:
                type: string
                example: "invalid login - bad password"
...